requirements.txt
Flask==3.1.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-openai==0.3.13
streamlit
requests==2.32.3
pyyaml==6.0.2
faiss-cpu==1.10.0